Create a new miyoo defconfig with video enabled
With this config, the first boot shows a NxHope splash screen, upside down. It apparently does the FS resizing and such in the background.
Then subsequent boots show the Miyoo Custom Firmware splash screen that's in 1.3.3, and sometimes linux can successfully take over.
Other times the screen hangs on the boot logo, but Linux is apparently running in the background.